Remove flights from account that have already departed
Since flights are run in a separate process than the account, the
list of flights in the account object will grow infinitely as the flight
can't remove itself once it is checked in (since it is not in the same
process).

Flights are removed if their departing time has already passed (Check
ins wouldn't matter after this point). A flag set by the flight can't be
used, again because it is in a separate process. Communicating between
processes would most likely be too complicated (at least the ways I've
tried).

This will be used with automatic scheduling. Since the script will run
indefinitely, we don't want the list of flights to grow infinitely.